requirements-tasks.txt
apscheduler>=3.10,<4.0
psycopg2-binary==2.9.9
requests==2.32.3
pydantic-settings==2.3.4
usaddress==0.5.10
